app/account/token.py
from datetime import datetime, timedelta
from jose import JWTError, jwt
from . import account_schemas

SECRET_KEY = "8e73bdfac849a294dc34181f77662d286bee3b53e80868056ce54e7c86f055a7"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 1440

def create_access_token(data: dict):
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt

def verify_token(token:str,credentials_exception):
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        email: str = payload.get("sub")
        if email is None:
            raise credentials_exception
        token_data = account_schemas.TokenData(email=email)
    except JWTError:
        raise credentials_exception